Move SnakeCaseNamingPolicy
