+++ tools/Titan.Tools.Packager/PreReqs/DownloadDXCCompiler.cs
@@ -93,25 +93,3 @@ internal class DownloadDXCCompiler
 }
 internal record GithubRelease(string Name, string TagName, DateTime PublishedAt, bool PreRelease, string Url, GithubAsset[]? Assets);
 internal record GithubAsset(string Name, string ContentType, string BrowserDownloadUrl, long Size);
-
-internal class SnakeCaseNamingPolicy : JsonNamingPolicy
-{
-    public override string ConvertName(string name)
-    {
-        Span<char> buffer = stackalloc char[name.Length * 2];
-        var charCount = 0;
-
-        for (var i = 0; i < name.Length; i++)
-        {
-            var character = name[i];
-            if (i != 0 && char.IsUpper(character))
-            {
-                buffer[charCount++] = '_';
-            }
-
-            buffer[charCount++] = char.ToLower(character);
-        }
-
-        return new string(buffer[..charCount]);
-    }
-}
